FS-5062 - Moving tests against generate_assessment_config out of test_integration and into test_generate_assessment_config unit test module
